bm-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
python-dotenv
httpx
oauthlib
authlib